Switch libmbfl to use size_t
Switch mbfl_string and related structures to use size_t lengths.

Quite likely that I broke some things along the way...

+++ ext/mbstring/libmbfl/mbfl/mbfl_convert.h
@@ -52,7 +52,7 @@ struct _mbfl_convert_filter {
 	const mbfl_encoding *to;
 	int illegal_mode;
 	int illegal_substchar;
-	int num_illegalchar;
+	size_t num_illegalchar;
 	void *opaque;
 };
